Better handling of throwing threads
Improved handling of the case where thread construction fails in the `dp::thread_pool` constructor. This could happen on a real machine due to limited resources. I simulated it in the unit tests by writing a class that throws based on a cut off probability. The thread class is now templated on the threadpool class too.